requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11.18
django-bootstrap3==11.0.0
django-markdownx==2.0.28
django-taggit==0.23.0
idna==2.5
Markdown==3.0.1
Pillow==5.4.1
pkg-resources==0.0.0
pytz==2018.9
requests==2.17.1
urllib3==1.21.1
